Change probability algorithm.
Still not correct (not sure what correct is) but I think it's simpler to reason about this way, and the results will always be 0 <= P <= 1.
Test some private methods as well - I had a math error I was having trouble finding and I hoped one of the methods had a bug.